Split test_pydecred.py into test_calc.py, text_txscript.py and test_vsp.py .
